@@ -1,12 +1,13 @@
 {
   "name": "virtual-sports-dot-be",
-  "version": "1.0.0",
+  "version": "1.1.0",
+  "type": "module",
   "description": "Virtual sports landing page based on an upgraded version of the 11ty-tailwind-alpinejs template: https://github.com/11ta/11ta-template v2.1",
   "author": "Thibault Fouquaert",
   "license": "GNU GPL V3",
   "engines": {
-    "npm": ">=8.0.0 <9.0.0",
-    "node": ">=16.0.0 <17.0.0"
+    "npm": ">=11.0.0 <12.0.0",
+    "node": ">=22.0.0 <23.0.0"
   },
   "bugs": {
     "url": "https://github.com/Thibaultfq/virtual-sports/issues"
@@ -29,33 +30,55 @@
     "dev:eleventy": "npx @11ty/eleventy --serve",
     "dev": "npm-run-all clean --parallel dev:*",
     "build:tailwind": "npx tailwindcss -i src/assets/css/tailwind.css -o dist/css/styles.css --minify",
-    "build:eleventy": "ELEVENTY_PRODUCTION=true npx @11ty/eleventy",
-    "build": "NODE_ENV=production npm-run-all clean build:*"
+    "build:eleventy": "npx @11ty/eleventy",
+    "build": "npm-run-all clean build:*"
   },
-  "devDependencies": {
-    "@11ty/eleventy": "^2.0.1",
-    "@11ty/eleventy-navigation": "^0.3.5",
+  "dependencies": {
+    "@11ty/eleventy": "^3.1.2",
+    "@11ty/eleventy-navigation": "^1.0.5",
     "@11ty/eleventy-plugin-rss": "^1.2.0",
-    "@tailwindcss/forms": "^0.5.10",
-    "@tailwindcss/typography": "^0.5.16",
-    "dotenv": "^16.4.7",
-    "glob": "^10.4.5",
-    "html-minifier": "^4.0.0",
-    "luxon": "^3.5.0",
+    "@11ty/eleventy-upgrade-help": "^3.0.2",
+    "@tailwindcss/forms": "^0.5.11",
+    "@tailwindcss/typography": "^0.5.19",
+    "dotenv": "^16.6.1",
+    "glob": "^10.5.0",
+    "html-minifier-terser": "^7.2.0",
+    "luxon": "^3.7.2",
     "markdown-it": "^13.0.2",
     "markdown-it-anchor": "^8.6.7",
     "markdown-it-emoji": "^2.0.2",
     "markdown-it-image-figures": "^2.1.1",
     "npm-run-all": "^4.1.5",
     "reading-time": "^1.5.0",
-    "rimraf": "^3.0.2",
+    "rimraf": "^6.1.2",
     "slugify": "^1.6.6",
     "svg-sprite": "^2.0.4",
-    "tailwindcss": "^3.4.17",
-    "terser": "^5.39.0",
-    "vinyl": "^3.0.0"
+    "tailwindcss": "^3.4.19",
+    "terser": "^5.44.1",
+    "vinyl": "^3.0.1"
   },
-  "dependencies": {
-    "debug": "^4.4.0"
+  "devDependencies": {
+    "@11ty/eleventy": "^3.1.2",
+    "@11ty/eleventy-navigation": "^1.0.5",
+    "@11ty/eleventy-plugin-rss": "^1.2.0",
+    "@11ty/eleventy-upgrade-help": "^3.0.2",
+    "@tailwindcss/forms": "^0.5.11",
+    "@tailwindcss/typography": "^0.5.19",
+    "debug": "^4.4.3",
+    "dotenv": "^16.6.1",
+    "glob": "^10.5.0",
+    "luxon": "^3.7.2",
+    "markdown-it": "^13.0.2",
+    "markdown-it-anchor": "^8.6.7",
+    "markdown-it-emoji": "^2.0.2",
+    "markdown-it-image-figures": "^2.1.1",
+    "npm-run-all": "^4.1.5",
+    "reading-time": "^1.5.0",
+    "rimraf": "^6.1.2",
+    "slugify": "^1.6.6",
+    "svg-sprite": "^2.0.4",
+    "tailwindcss": "^3.4.19",
+    "terser": "^5.45.0",
+    "vinyl": "^3.0.1"
   }
 }
