@@ -1,10 +1,10 @@
-require('dotenv').config()
+import 'dotenv/config'
 
 /**
- * Setup a current env and timestamp with timezone support 
+ * Setup a current env and timestamp with timezone support
  *  for use in footer and other non-content file locations
  */
-module.exports = {
+export default {
   env: process.env.ELEVENTY_ENV,
   timestamp: new Date(),
   timezone: process.env.TIMEZONE || 'UTC',
