@@ -1,7 +1,5 @@
-module.exports = {
-  init: (eleventyConfig) => {
-    return {
-      NODE_ENV: process.env.NODE_ENV?.trim() || 'development',
-    }
-  },
+export function init(eleventyConfig) {
+  return {
+    NODE_ENV: process.env.NODE_ENV?.trim() || 'development',
+  }
 }
